% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{constant_IntPPP_base_x_2_constant_IntPPP_base_y}
\alias{constant_IntPPP_base_x_2_constant_IntPPP_base_y}
\title{Convert constant Int$PPP series from one base year to another}
\usage{
constant_IntPPP_base_x_2_constant_IntPPP_base_y(gdp, base_x, base_y, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{base_x}{A double, base year of incoming constant gdp series}

\item{base_y}{A double, base year of outgoing constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert constant Int$PPP series from one base year to another
}
