% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{constant_IntPPP_base_x_2_constant_LCU_base_y}
\alias{constant_IntPPP_base_x_2_constant_LCU_base_y}
\title{Convert from constant Int$PPP in one base year to constant LCU of another}
\usage{
constant_IntPPP_base_x_2_constant_LCU_base_y(gdp, base_x, base_y, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{base_x}{A double, base year of incoming constant gdp series}

\item{base_y}{A double, base year of outgoing constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant Int$PPP in one base year to constant LCU of another
}
