% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{constant_IntPPP_base_x_2_current_USMER}
\alias{constant_IntPPP_base_x_2_current_USMER}
\title{Convert from constant Int$PPP base year x to current US$MER}
\usage{
constant_IntPPP_base_x_2_current_USMER(gdp, base_x, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{base_x}{A double, base year of incoming constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant Int$PPP base year x to current US$MER
}
