% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{constant_LCU_2_constant_USMER}
\alias{constant_LCU_2_constant_USMER}
\title{Convert from constant LCU to constant US$MER}
\usage{
constant_LCU_2_constant_USMER(gdp, base, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{base}{A double, the base year of incoming and outgoing constant GDP}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant LCU to constant US$MER
}
