% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{constant_LCU_base_x_2_current_USMER}
\alias{constant_LCU_base_x_2_current_USMER}
\title{Convert from constant LCU base x to current Int$PPP}
\usage{
constant_LCU_base_x_2_current_USMER(gdp, base_x, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{base_x}{A double, base year of incoming constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant LCU base x to current Int$PPP
}
