% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{constant_USMER_base_x_2_current_IntPPP}
\alias{constant_USMER_base_x_2_current_IntPPP}
\title{Convert from constant US$MER base year x to current Int$PPP}
\usage{
constant_USMER_base_x_2_current_IntPPP(gdp, base_x, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{base_x}{A double, base year of incoming constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant US$MER base year x to current Int$PPP
}
