% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGDPpc.R
\name{convertGDPpc}
\alias{convertGDPpc}
\title{Convert GDP per capita}
\usage{
convertGDPpc(
  gdppc,
  unit_in,
  unit_out,
  source = "wb_wdi",
  pop_source = source,
  verbose = FALSE
)
}
\arguments{
\item{gdppc}{A tibble with 3 columns "iso3c", "year" and "value", where value
is gdp per capita.}

\item{unit_in}{A string with the incoming GDP unit.}

\item{unit_out}{A string with the desired output GDP unit.}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{pop_source}{A string indicating the source database to use for
population numbers. Defaults to same source used for inflation and PPP
factors.}

\item{verbose}{TRUE or FALSE}
}
\value{
A tibble
}
\description{
Convert GDP per capita
}
