% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{current_IntPPP_2_constant_LCU_base_y}
\alias{current_IntPPP_2_constant_LCU_base_y}
\title{Convert from current Int$PPP to constant LCU base y}
\usage{
current_IntPPP_2_constant_LCU_base_y(gdp, base_y, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{base_y}{A double, base year of outgoing constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current Int$PPP to constant LCU base y
}
