% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{current_IntPPP_2_current_USMER}
\alias{current_IntPPP_2_current_USMER}
\title{Convert from current Int$PPP to current US$MER}
\usage{
current_IntPPP_2_current_USMER(gdp, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current Int$PPP to current US$MER
}
