% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{current_LCU_2_constant_LCU_base_2010}
\alias{current_LCU_2_constant_LCU_base_2010}
\title{Convert from current LCU to constant LCU base 2010}
\usage{
current_LCU_2_constant_LCU_base_2010(gdp, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current LCU to constant LCU base 2010
}
