% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{current_LCU_2_constant_LCU_base_y}
\alias{current_LCU_2_constant_LCU_base_y}
\title{Convert from current LCU to constant LCU base y}
\usage{
current_LCU_2_constant_LCU_base_y(gdp, base_y, source, linked = FALSE)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{base_y}{A double, base year of outgoing constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{linked}{A logical, use linked GDP deflator series or not. Defaults to
not.}
}
\description{
Convert from current LCU to constant LCU base y
}
