% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{current_LCU_2_current_IntPPP}
\alias{current_LCU_2_current_IntPPP}
\title{Convert from current LCU to current Int$PPP}
\usage{
current_LCU_2_current_IntPPP(gdp, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current LCU to current Int$PPP
}
