% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{current_LCU_2_current_USMER}
\alias{current_LCU_2_current_USMER}
\title{Convert from current LCU to current US$MER}
\usage{
current_LCU_2_current_USMER(gdp, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current LCU to current US$MER
}
