% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{current_USMER_2_constant_IntPPP_base_y}
\alias{current_USMER_2_constant_IntPPP_base_y}
\title{Convert from current US$MER to constant Int$PPP base y}
\usage{
current_USMER_2_constant_IntPPP_base_y(gdp, base_y, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{base_y}{A double, base year of outgoing constant gdp series}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current US$MER to constant Int$PPP base y
}
