% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_conversions.R
\name{current_USMER_2_current_IntPPP}
\alias{current_USMER_2_current_IntPPP}
\title{Convert from current US$MER to current Int$PPP}
\usage{
current_USMER_2_current_IntPPP(gdp, source)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current US$MER to current Int$PPP
}
