% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag2tibb.R
\name{mag2tibb}
\alias{mag2tibb}
\title{mag2tibb}
\usage{
mag2tibb(gdp)
}
\arguments{
\item{gdp}{A magclass object with gdp or gdppc data}
}
\value{
A tibble
}
\description{
Converts magclass object into tibble with colnames required
  for convertGDP.
}
