% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_2_tibb.R
\name{mag_2_tibb}
\alias{mag_2_tibb}
\title{mag_2_tibb}
\usage{
mag_2_tibb(gdp)
}
\arguments{
\item{gdp}{A magclass object with gdp or gdppc data}
}
\value{
A tibble
}
\description{
Converts magclass object into tibble with colnames required
for convertGDP.
}
