% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_user_input.R
\name{transform_internal}
\alias{transform_internal}
\title{Transform user input for package internal use}
\usage{
transform_internal(x, gdp, with_regions)
}
\arguments{
\item{x}{A tibble, the result of the internal conversion}

\item{gdp}{A tibble, data frame or magpie object, the latter of which
requires the \href{https://github.com/pik-piam/magclass}{magclass}
package to be installed. The data-frame needs to have at least 3 columns:
\itemize{
\item a character column with iso3c
(\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{wikipedia}) country codes,
\item a numeric column with years,
\item a numeric column named "value" with GDP values.
}}

\item{with_regions}{NULL or a data-frame. The data-frame should be "country to region
mapping": one column named "iso3c" with iso3c country codes, and one column named
"region" with region codes to which the countries belong. Any regions in the gdp
object will then be disaggregated according to the region mapping and weighed by the
GDP share of countries in that region in the year of the unit, converted on a country
level, and re-aggregated before being returned.}
}
\value{
x, with the same type and names of gdp
}
\description{
Transform user input for package internal use
}
