% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_user_input.R
\name{check_user_input}
\alias{check_user_input}
\title{Make sure that user input is valid}
\usage{
check_user_input(gdp, unit_in, unit_out, source, verbose)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{unit_in}{A string with the incoming GDP unit.}

\item{unit_out}{A string with the desired output GDP unit.}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{verbose}{TRUE or FALSE}
}
\value{
TRUE or an Error
}
\description{
check_user_input performs some checks on the function arguments.
}
