% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_user_input.R
\name{check_user_input}
\alias{check_user_input}
\title{Make sure that user input is valid}
\usage{
check_user_input(
  gdp,
  unit_in,
  unit_out,
  source,
  with_regions,
  replace_NAs,
  verbose
)
}
\arguments{
\item{gdp}{A tibble, data frame or magpie object, the latter of which
requires the \href{https://github.com/pik-piam/magclass}{magclass}
package to be installed. The data-frame needs to have at least 3 columns:
\itemize{
\item a character column with iso3c
(\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{wikipedia}) country codes,
\item a numeric column with years,
\item a numeric column named "value" with GDP values.
}}

\item{unit_in}{A string with the incoming GDP unit, one of:
\itemize{
\item "current LCU"
\item "current Int$PPP"
\item "current US$MER"
\item "constant YYYY LCU"
\item "constant YYYY Int$PPP"
\item "constant YYYY US$MER"
}
where YYYY should be replaced with a year e.g. "2010" or "2017".}

\item{unit_out}{A string with the incoming GDP unit, one of:
\itemize{
\item "current LCU"
\item "current Int$PPP"
\item "current US$MER"
\item "constant YYYY LCU"
\item "constant YYYY Int$PPP"
\item "constant YYYY US$MER"
}
where YYYY should be replaced with a year e.g. "2010" or "2017".}

\item{source}{Quosure of the source argument. See \code{\link[rlang:quosure]{rlang:quosure()}}.}

\item{with_regions}{NULL or a data-frame. The data-frame should be "country to region
mapping": one column named "iso3c" with iso3c country codes, and one column named
"region" with region codes to which the countries belong. Any regions in the gdp
object will then be disaggregated according to the region mapping and weighed by the
GDP share of countries in that region in the year of the unit, converted on a country
level, and re-aggregated before being returned.}

\item{replace_NAs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
NULL or 1 or "regional_average". Should countries for which
conversion factors are missing, have their NA-conversion factors replaced with 1 or
with a regional average? The default is no. If 1, then the conversion factors will be
set to 1, and essentially, no conversion will take place. If "regional_average" then,
the regional average of the region to which the country belongs to will be used. This
requires a region-mapping to be passed to the function, see the with_regions argument.}

\item{verbose}{TRUE or FALSE. A flag to turn verbosity on or off. Overrules
the GDPuc.verbose option, if it is set.}
}
\value{
TRUE or an Error
}
\description{
check_user_input performs some checks on the function arguments.
}
