% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{constant_USMER_2_constant_LCU}
\alias{constant_USMER_2_constant_LCU}
\title{Convert from constant US$MER to constant LCU}
\usage{
constant_USMER_2_constant_LCU(gdp, base, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{base}{A double, the base year of incoming and outgoing constant GDP}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from constant US$MER to constant LCU
}
