% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{constant_USMER_2_constant_LCU}
\alias{constant_USMER_2_constant_LCU}
\title{Convert from constant US$MER to constant LCU}
\usage{
constant_USMER_2_constant_LCU(gdp, base, source)
}
\arguments{
\item{gdp}{A tibble, data frame or magpie object, the latter of which
 requires the \href{https://github.com/pik-piam/magclass}{magclass}
 package to be installed. The data-frame needs to have at least 3 columns:
 \itemize{
   \item a character column with iso3c
   (\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{wikipedia}) country codes,
   \item a numeric column with years,
   \item a numeric column named "value" with GDP values.
}}

\item{base}{A double, the base year of incoming and outgoing constant GDP}

\item{source}{A string indicating the name of data frame to use for conversion
factors. Can be a custom data frame that exists in the calling environment, or
one of the package internal ones. Use \code{\link{print_source_info}}() to learn
about the available sources.}
}
\description{
Convert from constant US$MER to constant LCU
}
