% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGDP.R
\name{convertGDP}
\alias{convertGDP}
\title{Convert GDP time series data from one unit to another}
\usage{
convertGDP(
  gdp,
  unit_in,
  unit_out,
  source = "wb_wdi",
  with_regions = NULL,
  verbose = FALSE
)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{unit_in}{A string with the incoming GDP unit.}

\item{unit_out}{A string with the desired output GDP unit.}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{with_regions}{Data-frame or tibble with region mapping}

\item{verbose}{TRUE or FALSE}
}
\value{
A tibble with 3 columns
}
\description{
convertGDP converts GDP time series data from one unit to another, using GDP
deflators, market exchange rates (MERs) and purchasing power parity
conversion factors (PPPs) from `source`.
}
