% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGDP.R
\name{convertGDP}
\alias{convertGDP}
\title{Convert GDP time series data from one unit to another}
\usage{
convertGDP(gdp, unit_in, unit_out, source = "imf_weo", verbose = FALSE)
}
\arguments{
\item{gdp}{A tibble with 3 columns "iso3c", "year" and "value".}

\item{unit_in}{A string with the incoming GDP unit.}

\item{unit_out}{A string with the desired output GDP unit.}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{verbose}{TRUE or FALSE}
}
\value{
A tibble with 3 columns
}
\description{
# All on country level.
0# Converting current LCU (nominal) to constant LCU with base year x (real)
Need GDP deflator with base year x, unit = [.LCU/xLCU], often given in %
Divide nominal GDP by deflator in every year
1# Changing base year of constant LCU from x to y
Need either deflator with base year x, or current LCU series
With deflator: Multiply constant LCU with base year x [xLCU] with value in y
               of deflator with base year x [yLCU/xLCU]
With current series: Compute deflator (nominal divided by real), then follow above
2# Converting current LCU to current Int$MER or current Int$PPP
Need yearly Market exchange rates, or yearly PPPs
Multiply nominal GDP by chosen conversion factor
3# Converting constant LCU to constant Int$MER or constant Int$PPP, all same base year
Need base year Market exchange rates, or PPPs
Multiply real GDP by chosen conversion factor
3# Converting current LCU to constant Int$MER or constant Int$PPP, all same base year
Need base year Market exchange rates, or PPPs
Multiply real GDP by chosen conversion factor
4# Converting constant Int$MER to constant Int$PPP with base year x
Need MER and PPP of base year
5# Converting current Int$PPP to constant Int$PPP with base year x
Need yearly PPP conversion rate, GDP deflator with base year x
Divide current Int$PPP by PPP conversion rate, then follow 0, and 3
## Converting current LCU to Int$PPP Need yearly PPPs,
}
