% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elemental_conversions.R
\name{current_USMER_2_current_LCU}
\alias{current_USMER_2_current_LCU}
\title{Convert from current US$MER to current LCU}
\usage{
current_USMER_2_current_LCU(gdp, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Convert from current US$MER to current LCU
}
