% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_user_input.R
\name{disaggregate_regions}
\alias{disaggregate_regions}
\title{Take gdp data at regional level and disaggregate to country level}
\usage{
disaggregate_regions(gdp, with_regions, weight_unit, weight_year, source)
}
\arguments{
\item{gdp}{A tibble, data frame or magpie object, the latter of which
 requires the \href{https://github.com/pik-piam/magclass}{magclass}
 package to be installed. The data-frame needs to have at least 3 columns:
 \itemize{
   \item a character column with iso3c
   (\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{wikipedia}) country codes,
   \item a numeric column with years,
   \item a numeric column named "value" with GDP values.
}}

\item{with_regions}{NULL or a data-frame. The data-frame should be "country to region
mapping": one column named "iso3c" with iso3c country codes, and one column named
"region" with region codes to which the countries belong. Any regions in the gdp
object will then be disaggregated according to the region mapping and weighed by the
GDP share of countries in that region in the year of the unit, converted on a country
level, and re-aggregated before being returned.}

\item{weight_unit}{A string, either "PPP" or "MER}

\item{weight_year}{An integer equal to the base_x year}

\item{source}{A string indicating the name of data frame to use for conversion
factors. Can be a custom data frame that exists in the calling environment, or
one of the package internal ones. Use \code{\link{print_source_info}}() to learn
about the available sources.}
}
\value{
A tibble with the values gdp at country level
}
\description{
Take gdp data at regional level and disaggregate to country level
}
