% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_user_input.R
\name{disaggregate_regions}
\alias{disaggregate_regions}
\title{Take gdp data at regional level and dissagregate to country level}
\usage{
disaggregate_regions(gdp, with_regions, weight_unit, weight_year, source)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{with_regions}{NULL or a data-frame or tibble with region mapping}

\item{weight_unit}{A string, either "PPP" or "MER}

\item{weight_year}{An integer equal to the base_x year}

\item{source}{A string indicating the source database to use for conversion
factors.}
}
\value{
A tibble with the values gdp at country level
}
\description{
Take gdp data at regional level and dissagregate to country level
}
