% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDPuc-package.R
\docType{package}
\name{GDPuc-package}
\alias{GDPuc-package}
\alias{GDPuc}
\title{GDPuc package}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Provides A function to convert GDP time series from one unit to another.
All common GDP units are included, i.e. current and constant local currency
units (LCU), US$ via market exchange rates and international dollars via
PPP conversion factors. Conversion factors can easily be customized.
Conversion at a regional/aggregated level is also possible.
}
\keyword{internal}
