% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_source_info.R
\name{print_source_info}
\alias{print_source_info}
\title{Print information on sources to screen}
\usage{
print_source_info(source)
}
\arguments{
\item{source}{A string indicating the source database to use for conversion
factors.}
}
\description{
Print detailed information on sources to screen. Information includes the
name, origin, date, html-link and an associated note. Calling the function
without any argument will print information on all available sources.
}
