% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_source_info.R
\name{print_source_info}
\alias{print_source_info}
\title{Print information on sources}
\usage{
print_source_info(source)
}
\arguments{
\item{source}{Empty, or the name of one of the internal sources:
\enumerate{
\item "wb_wdi"
\item "wb_wdi_linked"
\item "wb_wdi_cpi"
}}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Print detailed information on conversion factor sources to the screen. Information includes the name, origin, date,
html-link and an associated note. Calling the function without any argument will print information on all
available sources.
}
\examples{
print_source_info()

}
