% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_user_input.R
\name{transform_internal}
\alias{transform_internal}
\title{Transform user input for package internal use}
\usage{
transform_internal(x, gdp, with_regions)
}
\arguments{
\item{x}{A tibble, the result of the internal conversion}

\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{with_regions}{Data-frame or tibble with region mapping}
}
\value{
x, with the same type and names of gdp
}
\description{
Transform user input for package internal use
}
