% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_user_input.R
\name{transform_user_input}
\alias{transform_user_input}
\title{Transform user input for package internal use}
\usage{
transform_user_input(gdp, unit_in, unit_out, source, with_regions)
}
\arguments{
\item{gdp}{A tibble or data-frame with at least 3 columns containing iso3c
country codes, years, and GDP values.}

\item{unit_in}{A string with the incoming GDP unit.}

\item{unit_out}{A string with the desired output GDP unit.}

\item{source}{A string indicating the source database to use for conversion
factors.}

\item{with_regions}{Data-frame or tibble with region mapping}
}
\value{
List
}
\description{
Transform user input for package internal use
}
